Refactors - migrated the AppBar.js content into the App.js and deleted the AppBar.js
