+++ client/src/components/CreateAccount.js
@@ -7,64 +7,64 @@ import Typography from '@material-ui/core/Typography';
 
 const Form = props => (
   <Grid container justify="center">
-  <form className="createAccountForm" >
-  <Typography align="center" variant="h4" color="inherit">
-            Create an Account
+    <form className="createAccountForm" >
+      <Typography align="center" variant="h4" color="inherit">
+        Create an Account
           </Typography>
-          <hr/>
-    <MenuItem>
-    Name:
+      <hr />
+      <MenuItem>
+        Name:
     <input
-      type="text"
-      name="username"
-      value={props.userVal}
-      onChange={props.handleChange}
-   
-    />
-    </MenuItem>
-    <MenuItem>
-    Password:
+          type="text"
+          name="username"
+          value={props.userVal}
+          onChange={props.handleChange}
+
+        />
+      </MenuItem>
+      <MenuItem>
+        Password:
     <input
-      type="password"
-      name="password"
-      value={props.passVal}
-      onChange={props.handleChange}
-    /> </MenuItem>
-    <MenuItem>
-    Nickname:
+          type="password"
+          name="password"
+          value={props.passVal}
+          onChange={props.handleChange}
+        /> </MenuItem>
+      <MenuItem>
+        Nickname:
     <input
-      type="text"
-      name="nickname"
-      value={props.nickVal}
-      onChange={props.handleChange}
-    /></MenuItem> <MenuItem>
-    Airport:
+          type="text"
+          name="nickname"
+          value={props.nickVal}
+          onChange={props.handleChange}
+        /></MenuItem> <MenuItem>
+        Airport:
     <input
-      type="text"
-      name="airport"
-      value={props.airVal}
-      onChange={props.handleChange}
-    /></MenuItem> 
-    <MenuItem>
-    Email:
+          type="text"
+          name="airport"
+          value={props.airVal}
+          onChange={props.handleChange}
+        /></MenuItem>
+      <MenuItem>
+        Email:
     <input
-      type="text"
-      name="email"
-      value={props.eVal}
-      onChange={props.handleChange}
-    /></MenuItem>
-    <MenuItem>
-    Budget:
+          type="text"
+          name="email"
+          value={props.eVal}
+          onChange={props.handleChange}
+        /></MenuItem>
+      <MenuItem>
+        Budget:
     <input
-      type="text"
-      name="budget"
-      value={props.budgetVal}
-      onChange={props.handleChange}
-    /></MenuItem>
-    <Button aligntItem="center" type="submit" onClick={props.submitUser}>
-      Create Account
+          type="text"
+          name="budget"
+          value={props.budgetVal}
+          onChange={props.handleChange}
+        /></MenuItem>
+      <Button aligntItem="center" type="submit" onClick={props.submitUser}>
+        Create Account
     </Button>
-  </form>
+    </form>
   </Grid>
 );
 
@@ -78,7 +78,7 @@ class AccountCreate extends React.Component {
     budget: ""
   };
 
-  componentDidMount() {}
+  componentDidMount() { }
   handleChange = e => {
     e.preventDefault();
     this.setState({
@@ -97,13 +97,13 @@ class AccountCreate extends React.Component {
     };
     console.log(userData)
     $.post("/api/users", userData)
-    .then(resp => {
-      console.log(resp);
-      alert("Thanks for Creating an account, Please Login to continue.");
-    })
-    .catch(err=>{
-      alert('Fill out the entire form!')
-    })
+      .then(resp => {
+        console.log(resp);
+        alert("Thanks for Creating an account, Please Login to continue.");
+      })
+      .catch(err => {
+        alert('Fill out the entire form!')
+      })
   };
   render() {
     return (
