+++ config/config.json
@@ -1,7 +1,7 @@
 {
     "development": {
       "username": "root",
-      "password": "georgia18",
+      "password": "Wayne0126!",
       "database": "clsr_db",
       "host": "127.0.0.1",
       "port": 3306,
